+++ lib/src/analyzer_task.dart
@@ -7,7 +7,7 @@ import 'package:grinder/grinder.dart';
 @Deprecated('should be provided by the Grinder package itself soon.')
 void analyzerTask(
     {List<String> files: const [], List<String> directories: const []}) {
-  final _files = []..addAll(files);
+  final _files = files.toList();
   directories.forEach((dir) => _files.addAll(new FileSet.fromDir(
           new io.Directory(dir),
           pattern: '*.dart',