+++ tool/grind.dart
@@ -1,3 +1,15 @@
 library bwu_grinder_tasks.tool.grind;
 
 export 'package:bwu_grinder_tasks/bwu_grinder_tasks.dart';
+/*
+export 'package:bwu_grinder_tasks/bwu_grinder_tasks.dart' hide main;
+import 'package:bwu_grinder_tasks/bwu_grinder_tasks.dart'
+    show testTask, testTaskImpl, grind;
+
+///
+void main([List<String> args]) {
+  testTask = ([String preset]) =>
+      testTaskImpl(preset, runPubServe: true, pubServeParameters: ['test']);
+  grind(args);
+}
+*/